// interfaces/Exercice.ts
export interface Exercice {
    id: number;
    name: string;
    repetitions: number
}

export interface allDataExercice {
    category_id: number;
    category_name: string;
    program_id: number;
    program_name: string;
    exercice_id: number;
    exercice_name: string;
    exercice_repetitions: number;
}